use class-based API, fix minor issues

+++ pipelines/matrix/src/matrix/pipelines/modelling/nodes.py
@@ -122,8 +122,8 @@ embeddings_schema = DataFrameSchema(
 
 class EmbeddingsWithPairsSchema(DataFrameModel):
     y: T.IntegerType() = Field()  # type: ignore
-    source_embedding: T.ArrayType(T.DoubleType()) = Field(nullable=False)  # type: ignore
-    target_embedding: T.ArrayType(T.DoubleType()) = Field(nullable=False)  # type: ignore
+    source_embedding: T.ArrayType(T.FloatType()) = Field(nullable=False)  # type: ignore
+    target_embedding: T.ArrayType(T.FloatType()) = Field(nullable=False)  # type: ignore
 
     class Config:
         strict = False
